+++ slash/runner.py
@@ -83,8 +83,8 @@ def _get_run_context_stack(test, test_iterator, fixture_scope_manager):
         stack.enter_context(_get_test_context(test))
         stack.enter_context(_get_test_hooks_context())
         stack.enter_context(_update_result_context())
-        stack.enter_context(_cleanup_context())
         stack.enter_context(_get_test_fixture_context(test, test_iterator, fixture_scope_manager))
+        stack.enter_context(_cleanup_context())
         stack.enter_context(handling_exceptions())
         yielded = True
         yield